split type defs into separate files
