+++ helpers/voting/formatVotes.ts
@@ -1,8 +1,7 @@
 import { formatFixed, parseFixed } from "@ethersproject/bignumber";
 import { BigNumber, BigNumberish } from "ethers";
-import { solidityKeccak256 } from "helpers";
+import { encryptMessage, getPrecisionForIdentifier, getRandomSignedInt, solidityKeccak256 } from "helpers";
 import { FormatVotesToCommit, VoteFormattedToCommitT, VoteT } from "types";
-import { encryptMessage, getPrecisionForIdentifier, getRandomSignedInt } from "helpers";
 
 function makeVoteHash(
   price: string,
