+++ types/index.ts
@@ -1,1 +1,7 @@
-export * from "./global";
+export * from "./delegation";
+export * from "./notifications";
+export * from "./queries";
+export * from "./staking";
+export * from "./ui";
+export * from "./user";
+export * from "./voting";
