+++ types/staking.ts
@@ -1,0 +1,9 @@
+import { BigNumber } from "ethers";
+
+export type StakerDetailsT = {
+  stakedBalance: BigNumber;
+  pendingUnstake: BigNumber;
+  canUnstakeTime: Date | undefined;
+  unstakeRequestTime: Date | undefined;
+  delegate: string;
+};
